requirements.txt
streamlit
langchain
langchain-openai
langgraph
pandas
numpy
pydantic>=2.0
pymupdf
python-dotenv
pyyaml
tqdm
typer
pillow
opencv-python-headless
